backend/requirements.txt
django==2.2.16
djangorestframework==3.12.4
django-filter==2.2.0
python-dotenv==0.21.0
pillow==9.4.0
psycopg2-binary==2.8.6
djoser==2.1.0
gunicorn==20.1.0
drf-extra-fields==3.4.1
isort==5.11.5
reportlab==3.6.12